Fixed RgbaUintBitCount VGlyph.Wrapper

+++ wrappers/csharp/VGlyph.Wrapper/Import/RgbaUintBitCount.cs
@@ -5,9 +5,9 @@ namespace VGlyph.Import
     [StructLayout(LayoutKind.Sequential)]
     internal struct RgbaUintBitCount
     {
-        public uint r;
-        public uint g;
-        public uint b;
-        public uint a;
+        public byte r;
+        public byte g;
+        public byte b;
+        public byte a;
     }
 }
